- polished UI (wip) - fixed Raycast collider issue - UI polish (finished) - added zoom functionality - finished timezone calculation and weather UI - included new font

+++ Assets/Scripts/Weather.cs
@@ -20,7 +20,7 @@ public class GetCityByName
 [System.Serializable]
 public class WeatherData
 {
-    public int timestamp;
+    public double timestamp;
     public TemperatureData temperature;
     public SummaryData summary;
 }
